requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.37.0
python-dotenv>=1.0.1
sqlmodel>=0.0.21
rank-bm25>=0.2.2
pydantic>=2.7.0
pydantic-settings>=2.3.4
httpx>=0.27.0
pymupdf>=1.24.10
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
nltk>=3.9
